Move env to commands.
