+++ hatch/commands/__init__.py
@@ -12,3 +12,4 @@ from hatch.commands.update import update
 from hatch.commands.clean import clean
 from hatch.commands.release import release
 from hatch.commands.python_path import python_path
+from hatch.commands.env import env
